requirements.txt
# ── Core ML ──────────────────────────────────────
torch>=2.2.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0

# ── LLM / LangChain ──────────────────────────────
langchain>=0.1.0
langchain-openai>=0.0.8
openai>=1.12.0

# ── API ───────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0

# ── Dashboard ─────────────────────────────────────
dash>=2.16.0
dash-bootstrap-components>=1.5.0
plotly>=5.20.0

# ── Streaming ─────────────────────────────────────
kafka-python>=2.0.2
redis>=5.0.3

# ── Persistence ───────────────────────────────────
influxdb-client>=1.42.0

# ── Dev / Testing ─────────────────────────────────
pytest>=8.1.0
pytest-cov>=5.0.0
httpx>=0.27.0           # required by FastAPI TestClient
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
